preparation for WSMAN support
CIMClient now forms an abstract base class for CIMXMLClient and for future WSMANClient.  WBEMConnectionBase is now base class for WBEMConnection; it handles backend creation and CIMClient* is now accessed via client().  WBEMConnectionBase is also used for selecting backend; depending if local_connection is set or "/wsman" path is used in URL.